Front back

+++ hybridpos-Frontend/src/app/mycomponents/admin-page/products/products.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, ChangeDetectorRef } from '@angular/core';
 import { FormsModule, NgForm } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 
@@ -12,8 +12,23 @@ import { ProductService } from '../../../services/product/product-service';
   styleUrl: './products.css',
 })
 export class Products {
-  selectedImage: File | null = null;
+  //product
   products: any[] = [];
+  selectedProduct: any = null;
+  //stock
+  selectedStockProduct: any = null;
+  stockAmount = 0;
+  //image
+  selectedImage: File | null = null;
+  selectedUpdateImage: File | null = null;
+  updateImagePreviewUrl: string | null = null;
+
+  updateProductData = {
+    barcode: '',
+    name: '',
+    purchasePrice: 0,
+    salePrice: 0,
+  };
 
   newProduct = {
     barcode: '',
@@ -30,7 +45,10 @@ export class Products {
 
   barcodeInput = '';
 
-  constructor(private productService: ProductService) {}
+  constructor(
+    private productService: ProductService,
+    private cdr: ChangeDetectorRef,
+  ) {}
 
   ngOnInit(): void {
     this.getProducts();
@@ -40,6 +58,8 @@ export class Products {
     this.productService.getAllProducts().subscribe({
       next: (res: any[]) => {
         this.products = res;
+        console.log('PRODUCTS:', res);
+        this.cdr.detectChanges();
       },
 
       error: (err) => {
@@ -95,12 +115,17 @@ export class Products {
   }
 
   deleteProduct(id: number): void {
+    const confirmed = confirm('Bu ürünü silmek istediğinize emin misiniz?');
+
+    if (!confirmed) {
+      return;
+    }
+
     this.productService.deleteProduct(id).subscribe({
       next: () => {
         alert('Ürün silindi.');
         this.getProducts();
       },
-
       error: (err) => {
         alert('Ürün silinemedi.');
         console.error(err);
@@ -122,6 +147,90 @@ export class Products {
       error: (err) => {
         alert('Fiyat güncellenemedi.');
         console.error(err);
+      },
+    });
+  }
+  openUpdateModal(product: any): void {
+    this.selectedProduct = product;
+
+    this.updateProductData = {
+      barcode: product.barcode,
+      name: product.name,
+      purchasePrice: product.purchasePrice,
+      salePrice: product.salePrice,
+    };
+    this.updateImagePreviewUrl = null;
+    this.selectedUpdateImage = null;
+  }
+  openStockModal(product: any): void {
+    this.selectedStockProduct = product;
+    this.stockAmount = 0;
+  }
+  increaseStock(): void {
+    if (!this.selectedStockProduct || this.stockAmount <= 0) {
+      return;
+    }
+
+    this.productService.increaseStock(this.selectedStockProduct.id, this.stockAmount).subscribe({
+      next: () => {
+        alert('Stok başarıyla güncellendi.');
+
+        this.selectedStockProduct = null;
+        this.stockAmount = 0;
+
+        this.getProducts();
+      },
+      error: (err) => {
+        console.error(err);
+        alert('Stok güncellenemedi.');
+      },
+    });
+  }
+
+  onUpdateImageSelected(event: any): void {
+    const file = event.target.files[0];
+
+    if (file) {
+      this.selectedUpdateImage = file;
+
+      this.updateImagePreviewUrl = URL.createObjectURL(file);
+    }
+  }
+
+  updateProduct(): void {
+    if (!this.selectedProduct) {
+      return;
+    }
+
+    const formData = new FormData();
+
+    formData.append(
+      'product',
+      new Blob([JSON.stringify(this.updateProductData)], { type: 'application/json' }),
+    );
+
+    if (this.selectedUpdateImage) {
+      formData.append('image', this.selectedUpdateImage);
+    }
+
+    this.productService.updateProduct(this.selectedProduct.id, formData).subscribe({
+      next: () => {
+        alert('Ürün başarıyla güncellendi.');
+
+        this.selectedProduct = null;
+        this.selectedUpdateImage = null;
+        this.updateImagePreviewUrl = null;
+
+        this.getProducts();
+      },
+      error: (err) => {
+        console.error(err);
+
+        if (err.error?.message) {
+          alert(err.error.message);
+        } else {
+          alert('Ürün güncellenemedi.');
+        }
       },
     });
   }
